Features/bills

+++ src/api/client.ts
@@ -37,7 +37,11 @@ client.interceptors.response.use(
     // Any status codes that falls outside the range of 2xx cause this function to trigger
     // Do something with response error
 
-    if (isAxiosError(error) && error.response?.status === 401) {
+    if (
+      isAxiosError(error) &&
+      error.response?.status === 401 &&
+      window.location.pathname !== Routes.Login
+    ) {
       // refresh token
       BrowserStorageService.remove(BrowserStorageKeys.AccessToken);
       window.location.href = `${window.location.origin}${Routes.Login}`;
